code/requirements.txt
pandas>=2.0.0,<3.0.0
numpy>=1.24.0,<2.0.0
scikit-learn>=1.3.0,<2.0.0
python-dateutil>=2.8.2,<3.0.0
pytz>=2023.3
jupyter>=1.0.0
matplotlib>=3.7.0,<4.0.0
seaborn>=0.12.0,<1.0.0